+++ tracing-appender/src/lib.rs
@@ -104,7 +104,11 @@
 //!     .init();
 //! # }
 //! ```
-#![doc(html_root_url = "https://docs.rs/tracing-appender/0.1.0")]
+#![doc(html_root_url = "https://docs.rs/tracing-appender/0.1.1")]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/tokio-rs/tracing/master/assets/logo.svg",
+    issue_tracker_base_url = "https://github.com/tokio-rs/tracing/issues/"
+)]
 #![warn(
     missing_debug_implementations,
     missing_docs,