+++ tracing-serde/src/lib.rs
@@ -1,3 +1,8 @@
+#![doc(html_root_url = "https://docs.rs/tracing-serde/0.1.1")]
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/tokio-rs/tracing/master/assets/logo.svg",
+    issue_tracker_base_url = "https://github.com/tokio-rs/tracing/issues/"
+)]
 #![warn(
     missing_debug_implementations,
     // missing_docs, // TODO: add documentation