+++ tracing-opentelemetry/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.6.0 (August 4, 2020)
+
+### Breaking Changes
+
+- Upgrade to `v0.7.0` of `opentelemetry` (#867)
+  For list of breaking changes in OpenTelemetry, see the
+  [v0.7.0 changelog](https://github.com/open-telemetry/opentelemetry-rust/blob/master/CHANGELOG.md#v070).
+
 # 0.5.0 (June 2, 2020)
 
 ### Added